Issue #4 : Expiry Time Calculated Incorrectly
Internal function to_datetime was expecting unix time in microseconds,
but the parser was providing a timestamp in milliseconds.

With this fix, we interpret the unix timestamp as UTC. The rdb file does
not store the timezone, so we cannot accurately convert it to the right
time. So, we cheat and use UTC always.