Fixing a endian-ness bug in read_ziplist_entry.
The Redis dump file is using little endian numbers everywhere except for the
length of ziplist entries of more than 16,383 bytes. That is a hard edge case
to trigger, because it requires increasing hash-max-ziplist-value, since the
default insures that this case will never happen. I've added a test to catch
the issue, but it requires setting hash-max-ziplist-value to over 20,000.