fix #106 - reinstate python 2.6 long type
looks like both the original code, and the one in the following commit:
d7965253799df9850b27a8c8106e89c303041e15
assumed a simple int type is 64bit (8 bytes), and a long int is 128bit (16 bytes).
this was wrong, but besides we now have the architecture flag that tells us the
size of an int on the target platform.