Warning users when python-lzf is not installed
Decompressing LZF strings from the rdb file takes a significant amount of time. The python-lzf package is a native c module that significantly speeds up decompressing. Without this package, rdbtools uses a pure-python method to decompress string, and this is excruciatingly slow.

With this pull request, we update the readme to also install python-lzf from pip. Additionally, if we detect that lzf is not installed, we warn the users and provide instructions to install it.